pyproject.toml
[tool.poetry]
name = "apis-in-nutshell"
version = "0.1.0"
description = ""
authors = ["Kathy <trankathy009@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
